Bending moment of 9 types of truck

When a truck travels on a bridges, we use static influence line to calculate unit load bending moment. We assume truck velocity is v(m/s), and Observation frequency is Hz,so We will get a sequence of bending moments within BridgeL/v time, every 1/Hz seconds, at the mid-span of bridge.

Usage

BendingMoment(Trucktype, BridgeL, v, Hz, NoiseLevel)

Arguments

Trucktype

9 types of trucks, with 1-9

BridgeL

Length of bridge (m)

v

velocity of truck (m/s)

Hz

Observation frequency.

NoiseLevel

Usually 5%, 10%, 15%, 20% noise to simulate the real data.

Value

A sequence of bending moments generated at mid-span of bridge, within BridgeL/v time, every 1/Hz seconds.

IL

Static Influence Line of Bridge

Details

We want to return a sequence of bending moments at BridgeL/v time, every 1/Hz seconds, at the mid-span of bridge. and in order to make data real, we add Noiselevel into data, and we have 9 types of default trucks, each has different alxes load and alxes spacings, which was indicated in internal data.

Author

Qian Shi

Examples

BendingMoment(1,100,50,50,0.15)
#>        T          M
#> 1   0.00    0.00000
#> 2   0.02    0.00000
#> 3   0.04    0.00000
#> 4   0.06  199.73809
#> 5   0.08  142.57805
#> 6   0.10   20.50042
#> 7   0.12  130.86774
#> 8   0.14  200.45002
#> 9   0.16  276.25852
#> 10  0.18  283.92726
#> 11  0.20  197.92811
#> 12  0.22  494.96122
#> 13  0.24  597.57179
#> 14  0.26  186.29342
#> 15  0.28  335.07257
#> 16  0.30  544.32364
#> 17  0.32  574.88087
#> 18  0.34  490.44979
#> 19  0.36  477.73674
#> 20  0.38  754.15325
#> 21  0.40  784.90768
#> 22  0.42  486.89280
#> 23  0.44  755.25395
#> 24  0.46  756.13799
#> 25  0.48  582.97684
#> 26  0.50  928.36227
#> 27  0.52 1015.70211
#> 28  0.54  961.44353
#> 29  0.56  924.26500
#> 30  0.58  690.21401
#> 31  0.60  865.78864
#> 32  0.62  934.60018
#> 33  0.64  972.33399
#> 34  0.66 1161.50890
#> 35  0.68  792.10260
#> 36  0.70 1437.29837
#> 37  0.72 1088.23762
#> 38  0.74 1122.98964
#> 39  0.76 1048.19032
#> 40  0.78 1366.59410
#> 41  0.80 1215.29796
#> 42  0.82 1365.27515
#> 43  0.84 1202.86358
#> 44  0.86 1460.57538
#> 45  0.88 1493.48344
#> 46  0.90 1407.30574
#> 47  0.92 1641.23549
#> 48  0.94 1383.97923
#> 49  0.96 1409.23229
#> 50  0.98 1590.14262
#> 51  1.00 1753.70389
#> 52  1.02 1522.87607
#> 53  1.04 1843.92333
#> 54  1.06 1541.03148
#> 55  1.08 1568.97643
#> 56  1.10 1538.48067
#> 57  1.12 1771.92499
#> 58  1.14 1297.27909
#> 59  1.16 1642.96210
#> 60  1.18 1295.34874
#> 61  1.20 1631.99866
#> 62  1.22 1370.74530
#> 63  1.24 1369.71972
#> 64  1.26 1176.76649
#> 65  1.28 1298.50488
#> 66  1.30 1245.43330
#> 67  1.32 1419.69315
#> 68  1.34 1105.27030
#> 69  1.36 1052.03677
#> 70  1.38 1130.28259
#> 71  1.40 1078.61575
#> 72  1.42  963.31545
#> 73  1.44 1106.81578
#> 74  1.46  828.34555
#> 75  1.48  862.71099
#> 76  1.50  961.53363
#> 77  1.52  813.45355
#> 78  1.54  782.02510
#> 79  1.56  832.42748
#> 80  1.58  747.24017
#> 81  1.60  587.77962
#> 82  1.62  760.35883
#> 83  1.64  506.67009
#> 84  1.66  309.63554
#> 85  1.68  853.68334
#> 86  1.70  592.52713
#> 87  1.72  435.59024
#> 88  1.74  576.23904
#> 89  1.76  273.30456
#> 90  1.78  465.98786
#> 91  1.80  361.37460
#> 92  1.82  164.00936
#> 93  1.84  118.15060
#> 94  1.86  431.19700
#> 95  1.88    0.00000
#> 96  1.90   11.05071
#> 97  1.92   64.27469
#> 98  1.94   16.84984
#> 99  1.96  161.71478
#> 100 1.98   31.06584